refactor: extract Header, Todos, Foter to separate components
