+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,39 @@
+export const Footer = () => {
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+      {/* Hide the footer if there are no todos */}
+
+      <span className="todo-count" data-cy="TodosCounter">
+        3 items left
+      </span>
+
+      {/* Active link should have the 'selected' class */}
+      <nav className="filter" data-cy="Filter">
+        <a href="#/" className="filter__link selected" data-cy="FilterLinkAll">
+          All
+        </a>
+
+        <a href="#/active" className="filter__link" data-cy="FilterLinkActive">
+          Active
+        </a>
+
+        <a
+          href="#/completed"
+          className="filter__link"
+          data-cy="FilterLinkCompleted"
+        >
+          Completed
+        </a>
+      </nav>
+
+      {/* this button should be disabled if there are no completed todos */}
+      <button
+        type="button"
+        className="todoapp__clear-completed"
+        data-cy="ClearCompletedButton"
+      >
+        Clear completed
+      </button>
+    </footer>
+  );
+};
