+++ src/components/Header/Header.tsx
@@ -1,0 +1,41 @@
+import { useState } from 'react';
+
+type Props = {
+  onAddTodo: (title: string) => void;
+};
+
+export const Header = ({ onAddTodo }: Props) => {
+  const [todoTitle, setTodoTitle] = useState('');
+
+  const handleSubmit = (event: React.FormEvent) => {
+    event.preventDefault();
+
+    if (todoTitle.trim()) {
+      onAddTodo(todoTitle);
+      setTodoTitle('');
+    }
+  };
+
+  return (
+    <header className="todoapp__header">
+      {/* this button should have `active` class only if all todos are completed */}
+      <button
+        type="button"
+        className="todoapp__toggle-all active"
+        data-cy="ToggleAllButton"
+      />
+
+      {/* Add a todo on form submit */}
+      <form onSubmit={handleSubmit}>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+          value={todoTitle}
+          onChange={event => setTodoTitle(event.target.value)}
+        />
+      </form>
+    </header>
+  );
+};
