+++ src/components/Todos/Todos.tsx
@@ -1,0 +1,124 @@
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  todos: Todo[];
+  onToggleTodoStatus: (id: number) => void;
+};
+
+export const Todos = ({ todos, onToggleTodoStatus }: Props) => {
+  return (
+    <section className="todoapp__main" data-cy="TodoList">
+      {/* This is a completed todo */}
+      {todos.map(todo => (
+        <div
+          key={todo.id}
+          data-cy="Todo"
+          className={`todo ${todo.completed ? 'completed' : ''}`}
+        >
+          {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+          <label className="todo__status-label" htmlFor={`todo-${todo.id}`}>
+            <input
+              data-cy="TodoStatus"
+              id={`todo-${todo.id}`}
+              type="checkbox"
+              className="todo__status"
+              checked={todo.completed}
+              onClick={() => onToggleTodoStatus(todo.id)}
+            />
+          </label>
+
+          <span data-cy="TodoTitle" className="todo__title">
+            {todo.title}
+          </span>
+          {/* Remove button appears only on hover */}
+          <button type="button" className="todo__remove" data-cy="TodoDelete">
+            ×
+          </button>
+        </div>
+      ))}
+      <div data-cy="Todo" className="todo completed">
+        {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+        <label className="todo__status-label">
+          <input
+            data-cy="TodoStatus"
+            type="checkbox"
+            className="todo__status"
+            checked
+          />
+        </label>
+
+        <span data-cy="TodoTitle" className="todo__title">
+          Completed Todo
+        </span>
+
+        {/* Remove button appears only on hover */}
+        <button type="button" className="todo__remove" data-cy="TodoDelete">
+          ×
+        </button>
+      </div>
+
+      {/* This todo is an active todo */}
+      <div data-cy="Todo" className="todo">
+        {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+        <label className="todo__status-label">
+          <input
+            data-cy="TodoStatus"
+            type="checkbox"
+            className="todo__status"
+          />
+        </label>
+
+        <span data-cy="TodoTitle" className="todo__title">
+          Not Completed Todo
+        </span>
+
+        <button type="button" className="todo__remove" data-cy="TodoDelete">
+          ×
+        </button>
+      </div>
+
+      {/* This todo is being edited */}
+      <div data-cy="Todo" className="todo">
+        {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+        <label className="todo__status-label">
+          <input
+            data-cy="TodoStatus"
+            type="checkbox"
+            className="todo__status"
+          />
+        </label>
+
+        {/* This form is shown instead of the title and remove button */}
+        <form>
+          <input
+            data-cy="TodoTitleField"
+            type="text"
+            className="todo__title-field"
+            placeholder="Empty todo will be deleted"
+            value="Todo is being edited now"
+          />
+        </form>
+      </div>
+
+      {/* This todo is in loadind state */}
+      <div data-cy="Todo" className="todo">
+        {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+        <label className="todo__status-label">
+          <input
+            data-cy="TodoStatus"
+            type="checkbox"
+            className="todo__status"
+          />
+        </label>
+
+        <span data-cy="TodoTitle" className="todo__title">
+          Todo is being saved now
+        </span>
+
+        <button type="button" className="todo__remove" data-cy="TodoDelete">
+          ×
+        </button>
+      </div>
+    </section>
+  );
+};
